refactor(agents): fold single-consumer subagent-spawn helpers back into callers
